+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [7.3.47] - 2026-01-21
+### Fixed
+- Fixed update time display in GroupPollCard [#5672](https://github.com/rokwire/illinois-app/issues/5672).
+- Fixed Large text issues in Assistant panel [#5610](https://github.com/rokwire/illinois-app/issues/5610).
+### Changed
+- Apply includeHidden param in groups / research project APIs [#5683](https://github.com/rokwire/illinois-app/issues/5683).
+
 ## [7.3.46] - 2026-01-15
 ### Fixed
 - Group Poll issues [#5519](https://github.com/rokwire/illinois-app/issues/5514).